Little bit of reorganization
