+++ fiberplane-charts/src/ACG/types.ts
@@ -160,7 +160,7 @@ export type Point<P> = {
 };
 
 /**
- * A bar to be rendered parallel to the Y axis.
+ * A rectangle to be rendered inside the chart.
  */
 export type Rectangle<P> = {
   xMin: number;
@@ -169,7 +169,7 @@ export type Rectangle<P> = {
   yMax: number;
 
   /**
-   * The source this bar was generated from.
+   * The source this rectangle was generated from.
    *
    * This would be a `Metric` if the chart was generated from `Timeseries`.
    */
