--- fiberplane-charts/src/CoreChart/ChartContent/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChartContent";
