+++ fiberplane-charts/src/CoreChart/ChartShape/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChartShape";
